+++ docs/en/changes/changes.md
@@ -57,6 +57,7 @@
 * Support HTTP URI formatting through `ai-pipeline` to do pattern recognition.
 * Add new HTTP URI grouping engine with benchmark. 
 * [Breaking Change] Use the new HTTP URI grouping engine to replace the old regex based mechanism.
+* Migrate from kubernetes-client/java to fabric8 client.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.